[COOP] ReportingObserver [1/M] Add WPT tests.
When an access is made in between two windows that would be put in
different browsing context group if COOP was enforced, we should
dispatch events to the ReportingObserver(s) and display error messages.

This patch adds the first two basic WPT tests. A few more are also
coming in follow-ups, in particular when the access is made from an
iframe.

Branch:
[1/M] This patch.

Bug:1111691
Change-Id: I38d7092c9edb664b9f2936c7a86f2c488fafde9c